Implemented phase shift on boundary + new query-flag system
Every "tag" in the original flag system has been replaced with a
cfg_state_t which contains the previous tag + a state bitmap. This will
tell you whether the associated config has smearing and phase shifts.
This will basically take care of Issue #40.

Also re-implemented my previous fix to the anti-period boundary
conditions (Issue #29), this was basically making the set_ud_phase() and
unset_ud_phase() functions also act on the boundary links so that we
didn't have to register it with a new tag. We use the new tagging system
to check for consistency with e.g. sw, u, grids etc.

Also updated the associated testing routines and made sure my unit tests
succeed.

Made the flags more consistent, added UD_IS_CLEAN

Replaced all the "is not phase shifted" tests with "is clean" tests so
that I test both phase shifting and smearing.

Also made sure that the smearing and phase shifting properly set the
flags for the boundary as well as these are in fact updated in tandem.

Some unit tests for the new flag system

Start fixing Issue #45. Would optimally have a couple more tests,
specially testing the other buffers (mostly focused on UDBUF), also need
to test some of the grid flags and events.