Prefixes all names with openqcd__
Made it so that if one defines the macro OPENQCD_INTERNAL one can
recover the shortened names as they originally were.

The extension unfortunately requires C99 to function due to the need for
variadic macros. This change will have to be benchmarked before we know
whether the solution can actually be used.

The names have in general also been prefixed with the module they come
from, so e.g. functions in forces will be calles "openqcd_forces__...".

TODO: Add the OPENQCD_INTERNAL flag to all the tests